+++ .test-infra/jenkins/job_PreCommit_Java_IOs.groovy
@@ -185,6 +185,9 @@ def additionalTasks = [
   snowflake: [
     ':sdks:java:io:snowflake:expansion-service:build',
   ],
+  jms: [
+    ':sdks:java:io:jms:integrationTest',
+  ],
 ]
 
 // In case the test suite name is different from the project folder name
@@ -194,6 +197,13 @@ def aliasMap = [
   'jdbc': 'JDBC',
   'hadoop-file-system': 'hadoop',
   'mongodb': 'MongoDb',
+]
+
+// In case the package name is different from the project folder name
+def packageNameMap = [
+  'amazon-web-services': 'aws',
+  'amazon-web-services2': 'aws2',
+  'hadoop-file-system': 'hadoop',
 ]
 
 ioModulesMap.forEach {cases, ioModules ->
@@ -211,7 +221,7 @@ ioModulesMap.forEach {cases, ioModules ->
     ]
     tasks.addAll(additionalTasks.get(it, []))
     def testName = aliasMap.get(it, it.capitalize())
-    String jacocoPattern = "**/org/apache/beam/sdk/io/${it}/**"
+    String jacocoPattern = "**/org/apache/beam/sdk/io/${packageNameMap.get(it,it)}/**"
     PrecommitJobBuilder builderSingle = new PrecommitJobBuilder(
         scope: this,
         nameBase: 'Java_' + testName + '_IO_Direct',