+++ sdks/python/apache_beam/internal/dill_pickler.py
@@ -46,6 +46,125 @@ import dill
 
 settings = {'dill_byref': None}
 
+if sys.version_info >= (3, 11) and dill.__version__ == "0.3.1.1":
+  # Let's make dill 0.3.1.1 support Python 3.11.
+
+  # The following function is based on 'save_code' from 'dill'
+  # Author: Mike McKerns (mmckerns @caltech and @uqfoundation)
+  # Copyright (c) 2008-2015 California Institute of Technology.
+  # Copyright (c) 2016-2023 The Uncertainty Quantification Foundation.
+  # License: 3-clause BSD.  The full license text is available at:
+  #  - https://github.com/uqfoundation/dill/blob/master/LICENSE
+
+  # The following function is also based on 'save_codeobject' from 'cloudpickle'
+  # Copyright (c) 2012, Regents of the University of California.
+  # Copyright (c) 2009 `PiCloud, Inc. <http://www.picloud.com>`_.
+  # License: 3-clause BSD.  The full license text is available at:
+  #  - https://github.com/cloudpipe/cloudpickle/blob/master/LICENSE
+
+  from types import CodeType
+
+  @dill.register(CodeType)
+  def save_code(pickler, obj):
+    if hasattr(obj, "co_endlinetable"):  # python 3.11a (20 args)
+      args = (
+          obj.co_argcount,
+          obj.co_posonlyargcount,
+          obj.co_kwonlyargcount,
+          obj.co_nlocals,
+          obj.co_stacksize,
+          obj.co_flags,
+          obj.co_code,
+          obj.co_consts,
+          obj.co_names,
+          obj.co_varnames,
+          obj.co_filename,
+          obj.co_name,
+          obj.co_qualname,
+          obj.co_firstlineno,
+          obj.co_linetable,
+          obj.co_endlinetable,
+          obj.co_columntable,
+          obj.co_exceptiontable,
+          obj.co_freevars,
+          obj.co_cellvars)
+    elif hasattr(obj, "co_exceptiontable"):  # python 3.11 (18 args)
+      args = (
+          obj.co_argcount,
+          obj.co_posonlyargcount,
+          obj.co_kwonlyargcount,
+          obj.co_nlocals,
+          obj.co_stacksize,
+          obj.co_flags,
+          obj.co_code,
+          obj.co_consts,
+          obj.co_names,
+          obj.co_varnames,
+          obj.co_filename,
+          obj.co_name,
+          obj.co_qualname,
+          obj.co_firstlineno,
+          obj.co_linetable,
+          obj.co_exceptiontable,
+          obj.co_freevars,
+          obj.co_cellvars)
+    elif hasattr(obj, "co_linetable"):  # python 3.10 (16 args)
+      args = (
+          obj.co_argcount,
+          obj.co_posonlyargcount,
+          obj.co_kwonlyargcount,
+          obj.co_nlocals,
+          obj.co_stacksize,
+          obj.co_flags,
+          obj.co_code,
+          obj.co_consts,
+          obj.co_names,
+          obj.co_varnames,
+          obj.co_filename,
+          obj.co_name,
+          obj.co_firstlineno,
+          obj.co_linetable,
+          obj.co_freevars,
+          obj.co_cellvars)
+    elif hasattr(obj, "co_posonlyargcount"):  # python 3.8 (16 args)
+      args = (
+          obj.co_argcount,
+          obj.co_posonlyargcount,
+          obj.co_kwonlyargcount,
+          obj.co_nlocals,
+          obj.co_stacksize,
+          obj.co_flags,
+          obj.co_code,
+          obj.co_consts,
+          obj.co_names,
+          obj.co_varnames,
+          obj.co_filename,
+          obj.co_name,
+          obj.co_firstlineno,
+          obj.co_lnotab,
+          obj.co_freevars,
+          obj.co_cellvars)
+    else:  # python 3.7 (15 args)
+      args = (
+          obj.co_argcount,
+          obj.co_kwonlyargcount,
+          obj.co_nlocals,
+          obj.co_stacksize,
+          obj.co_flags,
+          obj.co_code,
+          obj.co_consts,
+          obj.co_names,
+          obj.co_varnames,
+          obj.co_filename,
+          obj.co_name,
+          obj.co_firstlineno,
+          obj.co_lnotab,
+          obj.co_freevars,
+          obj.co_cellvars)
+    pickler.save_reduce(CodeType, args, obj=obj)
+
+  dill._dill.save_code = save_code
+
 
 class _NoOpContextManager(object):
   def __enter__(self):